+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.1.19]
+
+- Update Setup for functional tests
+
+- Add required changes to runtest for Ring 0 CI CD
+
 ## [0.1.18]
 
 - Handle exceptions inside interceptors